Convert MSP data frame into object of class Spectra

Convert msp data frame into object of class [Spectra()]

convertMsp2Spectra(msp)

Arguments

msp

data.frame that mimicks the .msp file format, see Details for further information.

Value

convertMsp2Spectra returns an object of class `Spectra`

Details

msp is a data frame of a .msp file, a typical data file for MS/MS libraries. The data frame has two columns and contains in the first column the entries "NAME:", "PRECURSORMZ:" (or "EXACTMASS:"), "RETENTIONTIME:", Num Peaks:" and information on fragments and peak areas/intensities and will extract the respective information in the second column.

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("convertMsp2Spectra", package = "MetCirc")
convertMsp2Spectra(msp = msp2spectra)
#> MSn data (Spectra) with 22 spectra in a MsBackendMemory backend:
#>       msLevel     rtime scanIndex
#>     <integer> <numeric> <integer>
#> 1           2        -1        NA
#> 2           2        -1        NA
#> 3           2        -1        NA
#> 4           2        -1        NA
#> 5           2        -1        NA
#> ...       ...       ...       ...
#> 18          2        -1        NA
#> 19          2        -1        NA
#> 20          2        -1        NA
#> 21          2        -1        NA
#> 22          2        -1        NA
#>  ... 17 more variables/columns.
#> Processing:
#>  Merge 2 Spectra into one [Wed Nov 30 11:18:26 2022]
#>  Merge 2 Spectra into one [Wed Nov 30 11:18:26 2022]