Create a link matrix

Create a link matrix which links every feature in similarity matrix with another.

createLink0df(similarityMatrix, sps, condition)

Arguments

similarityMatrix

matrix, a similarity matrix that contains the NDP similarity measure between all precursors in the data set

sps

Spectra object

condition

character, which conditions should be included?

Value

createLink0df returns a `matrix` that gives per each row information on linked features

Details

createLink0df creates a matrix from a similarity matrix which includes all connections between features in the similarity matrix, but exclude links which have a similarity of exactly 0.

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("spectra", package = "MetCirc")
data("similarityMat", package = "MetCirc")
link0df <- createLink0df(similarityMatrix = similarityMat,
    sps = sps_tissue, condition = c("SPL", "LIM", "ANT", "STY"))