Calculate similarity based on neutral losses

Calculate similarity based on neutral losses (NLS)

neutralloss(x, y, m = 0.5, n = 2, na.rm = TRUE, ...)

Arguments

x

Spectra object from Spectra containing intensity and m/z values, first MS/MS spectrum

y

Spectra object from Spectra containing intensity and m/z values, second MS/MS spectrum

m

numeric(1), exponent to calculate peak intensity-based weights

n

numeric(1), exponent to calculate m/z-based weights

na.rm

logical(1), if NA values should be removed

...

further arguments

Value

neutralloss returns a numeric similarity coefficient between 0 and 1

Details

Similarities of spectra based on neutral losses are calculated according to the following formula:

$$NLS = \frac{\sum(W_{S1, i} \cdot W_{S2, i}) ^ 2}{ \sum(W_{S1, i} ^ 2) \cdot \sum(W_{S2, i} ^ 2) }$$,

with \(W = [ peak intensity] ^{m} \cdot [ NL ]^n\) and \(NL = | m/z - precursor m/z |\). For further information see Li et al. (2015): Navigating natural variation in herbivory-induced secondary metabolism in coyote tobacco populations using MS/MS structural analysis. PNAS, E4147--E4155.

In here, the precursor m/z is taken by the m/z feature with the highest intensity.

neutralloss returns a numeric value ranging between 0 and 1, where 0 indicates no similarity between the two MS/MS features, while 1 indicates that the MS/MS features are identical. Prior to calculating $$W_{S1}$$ or $$W_{S2}$$, all intensity values are divided by the maximum intensity value.

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("spectra", package = "MetCirc")
Spectra::compareSpectra(sps_tissue[1:10], FUN = neutralloss, m = 0.5, n = 2)
#>               [,1]        [,2]        [,3]         [,4]         [,5]
#>  [1,] 1.0000000000 0.008478396 0.004164149 0.0484695131 0.0277748036
#>  [2,] 0.0084783956 1.000000000 0.289541770 0.1396731587 0.0012267464
#>  [3,] 0.0041641495 0.289541770 1.000000000 0.0524189976 0.0012581578
#>  [4,] 0.0484695131 0.139673159 0.052418998 1.0000000000 0.0008603832
#>  [5,] 0.0277748036 0.001226746 0.001258158 0.0008603832 1.0000000000
#>  [6,] 0.0015800500 0.009696469 0.006918308 0.0149447660 0.0403298823
#>  [7,] 0.0203651523 0.008660990 0.007077125 0.0019211038 0.0523635809
#>  [8,] 0.0018828751 0.004162380 0.006317503 0.0047400706 0.0810770346
#>  [9,] 0.0002955651 0.001372347 0.001689794 0.0085667616 0.0058115808
#> [10,] 0.0018171064 0.020847588 0.005229580 0.0116888876 0.0013743131
#>              [,6]        [,7]        [,8]         [,9]       [,10]
#>  [1,] 0.001580050 0.020365152 0.001882875 0.0002955651 0.001817106
#>  [2,] 0.009696469 0.008660990 0.004162380 0.0013723467 0.020847588
#>  [3,] 0.006918308 0.007077125 0.006317503 0.0016897939 0.005229580
#>  [4,] 0.014944766 0.001921104 0.004740071 0.0085667616 0.011688888
#>  [5,] 0.040329882 0.052363581 0.081077035 0.0058115808 0.001374313
#>  [6,] 1.000000000 0.018953115 0.008324715 0.0050980829 0.010025924
#>  [7,] 0.018953115 1.000000000 0.043348174 0.0057198110 0.004404181
#>  [8,] 0.008324715 0.043348174 1.000000000 0.0027337874 0.034361061
#>  [9,] 0.005098083 0.005719811 0.002733787 1.0000000000 0.053209777
#> [10,] 0.010025924 0.004404181 0.034361061 0.0532097766 1.000000000