Order columns and rows of a similarity matrix according to m/z, retention time and clustering

Internal function for shiny application. May also be used outside of shiny to reconstruct figures.

orderSimilarityMatrix(
  similarityMatrix,
  sps,
  type = c("retentionTime", "mz", "clustering"),
  group = FALSE
)

Arguments

similarityMatrix

matrix, similarityMatrix contains pair-wise similarity coefficients which give information about the similarity between precursors

sps

Spectra object containing spectra that are compared in similarityMatrix

type

character(1), one of "retentionTime", "mz", or "clustering"

group

logical(1), if TRUE group separated by "_" will be cleaved from rownames/colnames of similarityMatrix and matched against names of sps (sps$name), if FALSE rownames/colnames of similarityMatrix are taken as are and matched against names of sps (sps$name)

Value

matrix, orderSimilarityMatrix returns a similarity matrix with ordered rownames according to the character vector type

Details

orderSimilarityMatrix takes a similarity matrix, Spectra object (sps, containing information on m/z and retention time), and a character vector as arguments. It will then reorder rows and columns of the similarityMatrix object such, that it orders rows and columns of similarityMatrix according to m/z, retention time or clustering in each group.

orderSimilarityMatrix is employed in the shinyCircos function to create similarityMatrix objects which will allow to switch between different types of ordering in between groups (sectors) in the circos plot. It may be used as well externally, to reproduce plots outside of the reactive environment (see vignette for a workflow).

Author

Thomas Naake, thomasnaake@googlemail.com

Examples

data("spectra", package = "MetCirc")
similarityMat <- Spectra::compareSpectra(sps_tissue[1:10], 
    FUN = MsCoreUtils::ndotproduct, ppm = 10)
rownames(similarityMat) <- colnames(similarityMat) <- sps_tissue$name[1:10]

## order according to retention time 
orderSimilarityMatrix(similarityMatrix = similarityMat, 
    sps = sps_tissue, type = "retentionTime", group = FALSE)
#>           10         9         7         5         1         8         3
#> 10 1.0000000 0.5720148 0.5502293 0.5086787 0.1721383 0.5312776 0.3150198
#> 9  0.5720148 1.0000000 0.4940017 0.4581782 0.1367971 0.4968433 0.2905379
#> 7  0.5502293 0.4940017 1.0000000 0.5103709 0.2163514 0.5117921 0.3093910
#> 5  0.5086787 0.4581782 0.5103709 1.0000000 0.2252873 0.5052694 0.3307145
#> 1  0.1721383 0.1367971 0.2163514 0.2252873 1.0000000 0.1430105 0.2933251
#> 8  0.5312776 0.4968433 0.5117921 0.5052694 0.1430105 1.0000000 0.2809314
#> 3  0.3150198 0.2905379 0.3093910 0.3307145 0.2933251 0.2809314 1.0000000
#> 4  0.3293740 0.3610727 0.3740010 0.3935938 0.2559938 0.3519630 0.5362732
#> 6  0.4328983 0.4838513 0.4224726 0.4939638 0.1866679 0.5111684 0.3672317
#> 2  0.3105673 0.2784485 0.2955353 0.3627342 0.3270140 0.2933124 0.6828992
#>            4         6         2
#> 10 0.3293740 0.4328983 0.3105673
#> 9  0.3610727 0.4838513 0.2784485
#> 7  0.3740010 0.4224726 0.2955353
#> 5  0.3935938 0.4939638 0.3627342
#> 1  0.2559938 0.1866679 0.3270140
#> 8  0.3519630 0.5111684 0.2933124
#> 3  0.5362732 0.3672317 0.6828992
#> 4  1.0000000 0.4099733 0.5676489
#> 6  0.4099733 1.0000000 0.3548597
#> 2  0.5676489 0.3548597 1.0000000