Get MS/MS spectra that are present in condition

spectraCondition returns the names of Spectra that are present in condition, corresponding to the slot metadata.

spectraCondition(sps, condition)

Arguments

sps

Spectra object of Spectra package

condition

character, vector with conditions found as columns in the metadata slot

Value

list, named list with character vector as entries that contains the names of the MS/MS entries in spectra that are present in the conditon (tissues, stress conditions, time points, etc.)

Details

Helper function in createLink0df and shinyCircos.

Author

Thomas Naake thomasnaake@googlemail.com

Examples

data("spectra", package = "MetCirc")
MetCirc:::spectraCondition(sps = sps_tissue,
    condition = c("SPL", "LIM", "ANT", "STY"))
#> $SPL
#>  [1] "180" "181" "182" "183" "184" "185" "186" "187" "188" "189" "190" "191"
#> [13] "192" "193" "194" "195" "196" "197" "198" "199" "200" "201" "202" "232"
#> [25] "240" "241" "242" "243"
#> 
#> $LIM
#>  [1] "1"   "3"   "4"   "5"   "6"   "7"   "8"   "9"   "10"  "11"  "12"  "14" 
#> [13] "15"  "18"  "25"  "33"  "34"  "37"  "42"  "45"  "46"  "54"  "55"  "56" 
#> [25] "58"  "60"  "62"  "64"  "67"  "68"  "108" "109" "110" "111" "112" "113"
#> [37] "114" "115" "116" "117" "118" "119" "120" "121" "138" "157" "166" "167"
#> [49] "180" "181" "182" "183" "184" "185" "186" "187" "188" "189" "190" "191"
#> [61] "192" "193" "194" "195" "196" "197" "198" "199" "200" "201" "202" "207"
#> [73] "208" "211" "217" "219" "226" "227" "232" "233" "240" "242" "243" "255"
#> 
#> $ANT
#>   [1] "1"   "2"   "3"   "4"   "5"   "6"   "7"   "8"   "9"   "10"  "11"  "12" 
#>  [13] "14"  "15"  "16"  "17"  "18"  "24"  "25"  "33"  "34"  "37"  "38"  "40" 
#>  [25] "41"  "42"  "43"  "44"  "45"  "46"  "47"  "48"  "49"  "50"  "51"  "52" 
#>  [37] "53"  "54"  "55"  "56"  "58"  "59"  "60"  "61"  "62"  "63"  "65"  "66" 
#>  [49] "67"  "69"  "70"  "73"  "74"  "77"  "78"  "79"  "80"  "81"  "85"  "86" 
#>  [61] "87"  "89"  "90"  "91"  "92"  "94"  "95"  "96"  "97"  "98"  "99"  "100"
#>  [73] "102" "103" "104" "105" "106" "107" "108" "112" "142" "143" "144" "145"
#>  [85] "147" "148" "151" "152" "157" "158" "159" "160" "161" "162" "163" "164"
#>  [97] "166" "167" "168" "169" "171" "172" "175" "176" "177" "178" "179" "184"
#> [109] "185" "186" "187" "188" "191" "192" "194" "196" "199" "200" "201" "208"
#> [121] "211" "226" "227" "232" "233" "236" "255"
#> 
#> $STY
#>   [1] "19"  "20"  "21"  "22"  "23"  "24"  "25"  "26"  "27"  "28"  "29"  "30" 
#>  [13] "31"  "32"  "33"  "37"  "66"  "75"  "76"  "80"  "81"  "82"  "84"  "93" 
#>  [25] "99"  "101" "103" "104" "110" "113" "114" "115" "116" "117" "118" "119"
#>  [37] "120" "121" "123" "124" "125" "127" "128" "129" "130" "131" "132" "133"
#>  [49] "134" "135" "136" "137" "138" "139" "140" "141" "142" "143" "144" "145"
#>  [61] "146" "147" "148" "149" "151" "153" "154" "155" "156" "160" "165" "167"
#>  [73] "173" "174" "203" "204" "205" "206" "207" "208" "209" "210" "211" "212"
#>  [85] "213" "214" "215" "217" "218" "219" "220" "221" "222" "223" "224" "225"
#>  [97] "228" "229" "232" "233" "239" "244" "245" "246" "247" "248" "251" "253"
#> [109] "254" "255" "256" "258" "259"
#>